Creating StartPendingINspection style file
